ostree_git.bb: Bump forward to stable commit to pull in critical fixes
% git diff 13bcc49603b54f117c44e25dc2b457b9f25d9dc0 |diffstat
 README.md                          |    5 ++++-
 configure.ac                       |    4 ++--
 src/libostree/ostree-repo-commit.c |   14 ++++++++++++--
 3 files changed, 18 insertions(+), 5 deletions(-)

The main fix required was:
====
-          if (src_repo == OSTREE_REPO_MODE_BARE)
+          if (src_repo->mode == OSTREE_REPO_MODE_BARE)
====

The local pull operation would fail otherwise making it impossible
from a local disk repository.

Also changed in this commit was to drop the runtime dependency on
python.  There is absolutely no python inside the core OSTree
runtime.  Python is only used for the internal tests that are in the
OSTree source package, and these are not packaged.

[ Issue: LINUXEXEC-2447 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>