image_types_ota.bbclass: Always use core.add-remotes-config-dir false
The OSTree repo configuration files should always have all the
repository information stored in the main config file.  This allow for
simple copying of the config file when updating an A/B partition as
well as having all the core configuration in a single location vs
having it split out in to different config directories.

[ Issue: LINUXEXEC-2447 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>