ostree_upgrade_x86_64.sh: Recover all branches if the ostree repo is corrupt
Through testing of the upgrade from one branch to another, while also
corrupting the ostree repository on the upgrade partition, it was
found that there was no automated way to repair the branch that is not
the upgrade branch.  The test case was like this:

* boot off the A partition glibc-std
* Intentionally damage the ostree repo
* reboot
* Automated fall back will happen to the B partition
* From the B partition change the deployment branch to glibc-small
* Run ostree_upgrade.sh  (which upgrades the A partition)
* Reboot... And this point it will ultimately fall back to partition B
  because the original damage to the repository on the glibc-std branch was
  never repaired.

This patch adds in logic to ensure that the upgrade repository is both
repaired and passes a final fsck check, else the upgrade fails and
there is no state change.

[ Issue: LINUXEXEC-2447 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>