wr-ostree: Add single partition support to ostree_upgrade.sh for x86
The grub menus, ostree_upgrade.sh, and the network install all
properly support using either the A/B or none A/B upgrade scheme, with
rollback.

A regression had been created when the original A/B partition support
was added which made it impossible to update and offline partition and
manage the merge of the /etc directory with the ostree deploy
operation.  The original capability was restored and the A/B logic is
activated with an environment variable in the ostree_upgrade.sh
script.

At the point in time that the rootfs and /boot volume are capable of
being made read-only, the ostree_upgrade.sh now also supports
dynamically using an alternate read-write mount for the updates.

[ Issue: LIN1018-4534 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>