ostree-init: Add support for ARM for A/B and non A/B boot+rollback
The u-boot SD integration requires extra logic to deal with using a
single ostree partition.  Also added was the capability to use a
bootmenu with the boot.scr file when u-boot is compiled with the
options:

   CONFIG_CMD_BOOTMENU
   CONFIG_MENU

In the case that the u-boot doesn't have these options activated the
u-boot, the menu can be deactivated with: "setenv no_menu yes;
saveenv".  The bootmenu has a default internal timeout of 3 seconds
which can be changed with "setenv menutimeout XX" where XX is the
number seconds you want to wait, or set it to zero for infinite.

[ Issue: LINUXEXEC-3159 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>