ostree classes: Add generic ostree A/B, non A/B image generation
An end user can now customize the type of ostree image generated using
the following variables which you can set in the local.conf.

  OSTREE_USE_AB ?= "1"    -- Use the A/B partition or single partition image
  OSTREE_REMOTE_URL ?= "" -- URL of OSTree repository for upgrade, if you
                             bake it in, you can run ostree_upgrade.sh
			     each time you you rebuild and publish the updates

The wr-ostree layer will now use a .wks.in template with variable
substitutions so it works with all the image recipes.  There is one
.wks.in file for the A/B partition scheme and one for the non-ab
partition scheme.  The boot loader plays a role in the layout and file
insertion so there are 2 templates for each type of bootloader.  Grub
for example uses ostree-grub-ab.wks.in or ostree-grub-noab.wks.in.

The variables shown below are WIC specific image parameters which an
end user can change in the local.conf file.  By default they are all
empty.  An example might be as follows:

    OSTREE_WKS_EFI_SIZE = "--size=32M --overhead-factor 1"
    OSTREE_WKS_BOOT_SIZE = "--size=300M --overhead-factor 1"
    OSTREE_WKS_ROOT_SIZE = "--size=1024M --overhead-factor 1"
    OSTREE_WKS_FLUX_SIZE = "--size=2G --overhead-factor 1"

If using the variables above in the local.conf it has the following
impact.  The EFI partition will be 32 MB (or 32 MB fat partition in
the case of the u-boot integration).  The /boot volume will be 300MB,
the /root volume would be 1GB and the /var volume would be 2GB.  It is
important to leave enough space free in your /root volume because that
is where the ostree database lives and it should be sized to
accommodate how ever many ostree branches you need to store.

[ Issue: LINUXEXEC-3159 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>