image_types_ostree.bbclass: Add gpg 1.x support and fix pinentry for gpg 2.x
The gpgme library will only call "gpg" and search the path for the
binary.  In order to get ostree signing to call the correct binary
defined by GPG_BIN (such as /usr/bin/gpg2 or /usr/bin/gpg), a wrapper
script is required.

In the case of gpg 1.x the passphrase must be added as an argument in
order to be able to properly use a passphrase with a GPG key.  This is
taken care of in the generated wrapper script, which automatically
removed after it is used.

For the gpg 2.x support allow-loopback-pinentry must be set in the
gpg-agent.conf file in order to allow the passphrase to get cached
into memory.

[ Issue: LIN1018-4669 ]
[ Issue: LIN1018-4671 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>