wr-ostree configuration:  Add OSTREE_WKS_FILE definition
The feature/ostree template previously set the definition of the
WKS_FILE, the defaults for which template to use really belong to the
layer.  The feature/ostree template is intended to have only what is
needed to activate the layer.

Now an end user can change the ostree specific wks file with the
variable OSTREE_WKS_FILE in the local.conf, should it need to be
customized.

[ Issue: LINUXEXEC-3159 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>