u-boot-env: Add support for boot.scr booting with u-env
The prior version of boot.scr and the ostree integration used a
uEnv.txt file that is controlled by ostree and is not signed and the
possibility exists for tampering.  The u-boot is often not compiled
with the "env import" command.

The new default is to use the OSTREE_BOOTSCR = "fs_links", which will
make use of the symlinks that are managed via ostree to correctly
mount the boot/root partitions.  There is no speed trade off, and from
a functionality stand point most production deployment would want a
locked down uboot environment.

[ Issue: LINUXEXEC-3159 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>