ostree /init: Change default for OSTREE_SKIP_BOOT_DIFF=2
The type of deployment checks at boot time are decided by what is
local.conf or at install time with the network installer.  This value
can also be changed at run time.

The OSTREE_SKIP_BOOT_DIFF can be one of the following:
   0 = Perform a full fsck and boot diff
   1 = Skip the FSCK check
   2 = Skip the FSCK and the boot diff

The new default is: OSTREE_SKIP_BOOT_DIFF ??= "2"

At install time you can use the kernel argument:
    instsbd=0
  or
    instsbd=1

At runtime you can use:
    ostree config set upgrade.skip-boot-diff 0
  or
    ostree config set upgrade.skip-boot-diff 1

The OS will start significantly slower with the full fsck if it is a
large file system.  For a small file system it might only take a few
seconds extra.  The end user should decide the best deployment option.

Having this check at boot time allows for the highest
degree of integrity of the file system.  All the files in the local
ostree repository get checked against the signed ostree checksums for
the fsck phase.  For the boot diff phase all the files in the
deployment area are scanned and signature checked against contents of
the local ostree repository.

[ Issue: LINUXEXEC-3159 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>