ostree boot /home:  Fix the /home mount to instantiate in /var
It should be possible to reset /etc or /var and have any of the
template files that are supposed to get installed in /home get
installed again.

The ostree-upgrade-mgr also should not get installed into the
intramfs.  The dependency was incorrect, which has also been fixed.

The ostree bbclass which assembles the image moves any files found in
/home to /usr/homdirs/home.  The first boot will mirror anything in
the read-only copy into the newly auto created directory in /var/home
which is linked to /home, like any other system using libostree.

[ Issue: LIN1018-4816 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>