image_types_ostree.bbclass ostree: Use gnupg-native by default
It seems there is a myriad of problems with the signing tools when
relying upon the host version of the gpg binary.  This can be resolved
by depending on and using the gnupg-native package.

If GPG_BIN is set in the local.conf to an absolute path for the gpg
binary, there is still the possibility that this all works, but it is
not deterministic when building across different hosts, as the version
of gpgme might be different.

[ Issue: LIN1018-4765 ]
[ Issue: LIN1018-4835 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>