wr-ostree: get the valid value of the booting partition flag
Issue: LINUXEXEC-4069

Base on ostree architecture, there are 2 sets of system, called
A and B. And there is a flag that is used to indicate which system
boots up after upgrading operation.
Command setexpr.l is used to get the flag, which is a 4 bytes integer.
But on arm64 platform, this command get 8 bytes integer. The
high 4 bytes are random value, and cause later code logic is not
correct.
Therefore, it is need to do AND operation with 0xffffffff to get the
expected value.

Signed-off-by: Meng Li <Meng.Li@windriver.com>