init-ostree.sh: Improve speed of booting with luks volumes
It was taking and additional 10-30 seconds to boot with luks volumes
due to changes with blkid.  This commit removes the redundant checks
that use blkid and instead lets the init.luks-ostree script set up the
devices, since it is already performing doing the correct actions.

A shell should only be offered from a boot failure if running in debug
mode with "debugfatal" specified as a kernel boot argument by the boot
loader.

[ Issue: LINUXEXEC-5303 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>