init-ostree-install.sh: Update network installer to support encrypted volumes
The network installer had 3 new arguments added:

 LUKS=0                                - Do not create encrypted volumes
 LUKS=1                                - Encrypt var volume (requires TPM)
 LUKS=2                                - Encrypt var and and root volumes (requires TPM)

The default mode when LUKS is not set is LUKS=0.  It should be noted
that this only works on gpt volumes and requires a board with a TPM.

The "instsf=1" option was also fixed for the GPT disks.  This will
allow the installer to skip formatting the EFI volume when performing
the install.

[ Issue: LINUXEXEC-5303 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>