u-boot-uenv: Add support for device from ram
The raspberry pi uses the fdt_addr as a location in ram where the
device tree is pass from the firmware.  This means the variables need
to be separated.  The loadaddr is safe to use a place for the
temporary comparisons for the boot fallback logic, instead of using
the fdt_addr.

A new variable called use_fdtdtb was added to specify if the boot.scr
should make use of the ftd file provided by the boot code.

[ Issue: LINUXEXEC-5290 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>