ostree: Add feature OSTREE_ETC_MERGE_DIR=none
This feature will allow you to perform a deploy and restore the
contents of the /etc directory at deploy time.

It would be run like this:

OSTREE_ETC_MERGE_DIR=none ostree admin deploy ...deploy args...

This is extremely useful for resetting a device in a single step or
performing an upgrade where you want the factory settings at the next
boot cycle.

[ Issue: LINUXEXEC-5290 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>