ostree_upgrade.sh and ostree-init: Add the ability to local reset /var and /etc
The /var volume can now be restored to the factory settings (meaning
it is erased or formatted) as long as the the local.conf did not have
'OSTREE_ALLOW_RM_VAR = "0"'.

The ostree_upgrade.sh now has some arguments to make it a nicer
workflow so you that you do not have to run more than one command.
The following are the new arguments:

  -b   reboot after completion
  -e   Erase the /var volume on the next reboot
  -E   FORMAT the /var volume when on a separate partition on the next reboot
  -f   Force /etc to be entirely reset to the initial deploy state
  -r   Redeploy the current branch without doing a network pull
  -s   Skip the fsck integrity checks

  -F   Local Factory reset, uses -b -e -f -r -s
  -U   Factory upgrade reset, uses -b -e -f -s

[ Issue: LINUXEXEC-5290 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>