bootfs.sh: Add support for grub x86 bootstrap install
Now that instsf=1 is fixed for gpt partitions, it is possible to
create mini boot strap images that use grub or the grub network
install.

The bootfs directory can be exported for use with a tftp server to
network start a device with ipxe capabilities.

The script has also been changed to directly use the host tools from
the built image, as well as passing the LUKS=1 when it is a secure
image.  To additionally select the LUKS=2, which will use file system
encryption for the rootfs volume you would run:

   bootfs.sh -a LUKS=2

QEMU can demonstrate net starting by running:

 bitbake ovmf
 ../layers/wr-ostree/scripts/bootfs.sh -n
 qemu-img create -f raw img 10G
 dd if=ustart.img of=img conv=notrunc
 cp tmp*/deploy/images/*/ovmf.qcow2 OVMF_CODE.fd
 cp tmp*/deploy/images/*/ovmf.vars.qcow2 OVMF_VARS.fd
 qemu-system-x86_64 --enable-kvm --nographic -vnc :5 \
    -drive file=OVMF_CODE.fd,format=raw,readonly,if=pflash,unit=0 \
    -drive if=pflash,format=raw,unit=1,file=OVMF_VARS.fd \
    -drive file=img,if=virtio,format=raw -m 2048 \
    -serial mon:stdio \
    -netdev type=user,id=h1,tftp=bootfs,bootfile=EFI/BOOT/bootx64.efi \
    -device e1000,netdev=h1,mac=00:55:55:01:01:01 -usbdevice tablet \
    -cpu qemu64,+ssse3,+sse4.1,+sse4.2,+x2apic,rdrand -device virtio-rng-pci

[ Issue: LINUXEXEC-5304 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>