init-ostree-install.sh, bootfs.sh: Add support for local install generation
Now an image can generated with the OSTree repository builtin to the
EFI partition to execute a local install, while allowing the upgrade
to occur with what ever the OSTREE_REMOTE_URL is set to in the
local.conf.

It is also possible to use the "bootfs.sh -l OTHER_REPO_DIR", in case
you wanted to generate a repository that had only a single branch so
that it is more compact.

The local install image can be tested as follows:

     # Build the binaries
     # Don't forget to copy the ostree_repo somewhere
     # And set OSTREE_REMOTE_URL properly
     bitbake core-image-minimal ovmf

     # Build a boot disk called "img"
     ../layers/wr-ostree/scripts/bootfs.sh -n -L -s 512
     qemu-img create -f raw img 10G
     dd if=ustart.img of=img conv=notrunc

     # Now create the network install area
     ../layers/wr-ostree/scripts/bootfs.sh -w

     # Now boot QEMU with the ovmf firmware and network
     cp tmp*/deploy/images/*/ovmf.qcow2 OVMF_CODE.fd
     cp tmp*/deploy/images/*/ovmf.vars.qcow2 OVMF_VARS.fd
     qemu-system-x86_64 --enable-kvm --nographic -vnc :5 \
        -drive file=OVMF_CODE.fd,format=raw,readonly,if=pflash,unit=0 \
        -drive if=pflash,format=raw,unit=1,file=OVMF_VARS.fd \
        -drive file=img,if=virtio,format=raw -m 2048 \
        -serial mon:stdio \
        -netdev type=user,id=h1,tftp=bootfs,bootfile=EFI/BOOT/bootx64.efi \
        -device e1000,netdev=h1,mac=00:55:55:01:01:01 -usbdevice tablet \
        -cpu qemu64,+ssse3,+sse4.1,+sse4.2,+x2apic,rdrand -device virtio-rng-pci

For added excitement, after the device fully boots, you can destroy it
and watch it start off the network and fix itself.  You can do try it
as follows:

   # login as root
   dd if=/dev/zero of=/dev/vda bs=1M count=10
   reboot

The QEMU device should now load the contents of the tftp area and
fully re-install so long as the OSTREE_REMOTE_URL and repository are
served up properly.

[ Issue: LINUXEXEC-5304 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>