boot.scr, bootfs.sh: Add proper -a and -L support for bootfs.sh for u-boot
The extra argument handling for bootargs needed to be adjusted so that
a local install would work properly when generating it for a target
that uses u-boot.

The test case for the Raspberry Pi4 looks like this:

  bitbake core-image-minimal
  ../layers/wr-ostree/scripts/bootfs.sh -L -s 0
  # Flash that to a mmc device

The target should auto install, and you try something very destructive
to force a re-install, by deleting the extended partition.

  # Login as root
  fdisk /dev/mmcblk0
  d
  2
  w
  reboot

[ Issue: LINUXEXEC-5304 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>