init-ostree-install.sh: Ignore kernel arguments without the "=" sign
The kernel will pass any kernel arguments it does not process which do
not contain an "=" character as arguments to the install init script.

This will result in a kernel crash.  To be more graceful the install
script should continue to process the install.  If an end user wants
to debug what is happening in the installer, the "instsh=2" should be
passed as a kernel parameter.

[ Issue: LINCD-553 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>