ostree u-boot wic images: only partition 1 should be marked as active
Some configurations of u-boot such as the raspberry pi4 scan only the
active partition for the boot.scr file.  The /var partition was marked
as active in the wks template which causes the boot.scr search to
fail.  Only the first FAT partition should be marked as bootable in
the image.  That patch removes the active partition label from the
/var partition.

[ Issue: LIN1019-3813 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>