init-ostree-install.sh: installing from one device to another fails
If a disk already had partitions it will fail to delete the old ones
and the install will fail, because the image may be left with
just a single partition, or some cases multiple partitions.

The changes for deleted partitions must be committed and the
collection string for the partition operations cleared when using
instsf=1.

[ Issue: LINCD-671 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>