bootfs.sh, grub.cfg: Add auto install support option for x86
When using "bootfs.sh -l 0", it should create an image which auto
installs and it should be possible to re-install the device from the
grub menu at a later point, like you can with the u-boot version.

This commit adds in the creation of a igrub.cfg which can extend the
grub.cfg and will still be signature checked if using the secure boot
functions to allow for the auto re-install of a device when the kernel
and fallback kernel cannot be found.

[ Issue: LINCD-817 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>