image_types_ota.bbclass: Fix wic generated images using the A/B layout
The no_ab file which indicates that A/B disk layout is not in use was
getting placed into the image due the install script checking the
wrong variable.  This causes the fall back boot image to fail to load
because the boot.scr script will look in the wrong location.

[ Issue: LINCD-759 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>