bootfs.sh: Use mktemp to create gpg home for signing
On some host OS's the gpgconf does not make use of the /run/user/*
area for the unix sockets which have a hard limit of 108 characters
for the path.  Using mktemp uses a shortened path for the signing.

Also fix up the grub efi loader detection to look for just single file
either grubx64.efi grub-efi-grubx64.efi.

[ Issue: LINCD-958 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>