ostree-initramfs: /init needs to wait for blkid to complete
If using a slow USB device or MMC card the initramfs init script needs
to loop and wait for blkid to complete and return a partition label
for the boot volume.

When using an encrypted volume the possibility existed to call the
luks setup prior to the USB device being available for use leading to
a boot failure.  There was a retry loop to mount the root volume which
allowed 1/2 of the partitions to succeed in this corner case, but
still leads to a boot failure.

Each of the wait loops has been improved to utilize the max failure
delay and then call fatal which will allow the boot loader to try a
fallback entry.

Also fixed in this commit was to remove the shell call and replace it
with a call to fatal in the case that the ostree repository is damaged
as well as eliminating some debug output.

[ Issue: LINCD-833 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>