bootfs.sh, init-ostree-install.sh: Fix install arguments ordering
In the bootfs.sh the install arguments are not treated the same on the
x86 platform as ARM platform.  This patch will move all the appended
arguments to take priority.  That means what ever argument comes last
will win.  For a situation like "instab=0 instab=1", the instab=1 will
be the winner.

[ Issue: LINCD-1128 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>