init-ostree.sh/init-ostree-install.sh: preserve xattr attributes
Issue: LIN1019-4225

The original source files are all labeled with xattr attributes in
cgl+ostree image, when install it with "cp" to fluxdata partition,
xattr will be lost, read or write files without xattr will be denied
if selinux enabled. because the old version "cp" of initramfs do not
support xattr management, consider the size of initramfs, replace
"cp" with "tar" to avoid xattr missing issue.

Because the kernel is selinux enabled, operation on xattr will be
denied by kernel, while initramfs is not support selinux, so we
need disable kernel selinux to make tar have the permission to
manage xattr, after that, a reboot is necessary to avoid system
boot into a selinux disabled environment.

Signed-off-by: Liwei Song <liwei.song@windriver.com>