init-ostree.sh: move /run over to the new rootfs
/run was missing. This was causing boot failures on systems built using
the feature/encrypted-storage flag.

The udev database is stored in /run/udev/data. This database can be
repopulated in part but not entirely. It is especially important in
the case of the 10-dm.rules udev rules. Excerpt:

    If udev is used in initrd, we require the udev init script to not
    remove the existing udev database so we can reuse the information
    stored at the time of device activation in the initrd.

By moving the content of /run the boot process now completes without
errors.

Signed-off-by: Marc Ferland <ferlandm@amotus.ca>