mttyexec: Add tty input/output mux for initramfs
The mttyexec function can split the /dev/console device and redirect
input and output against all the console's specified by
console=... kernel arguments.

This provides a better experience to the end user to see the results
of a deployment on a graphical and or serial console.  It will also
allow in the future for the end user to enter in any provisioning
information such as WiFi settings or a one time password on what ever
console is available.

[ Issue: LINUXEXEC-8353 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>