bootloader console interface: Parameterize console= kernel argument
Switch all the boot loader configurations and bootfs.sh configuration
to allow the specification of the console via the OSTREE_CONSOLE
variable.

The mttyexec will redirect input and output to the install script in
the initramfs to all consoles.  Previously the console service was
split and only available to either the graphics or serial console.
Now it is available to up to 10 consoles, but most targets will only
have 1 or 2.

In the case that no console is specified, it will automatically fall
back to using tty0 as defined by the kernel interface.

[ Issue: LINUXEXEC-8353 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>