init-ostree-install.sh: Add the ability to provision with wifi
When the initramfs has the wpa_supplicant and the wifi drivers
available it is possible to ask for the wifi configuration on the
console at install time.

Documentation will be added a later point as more code is still
required in order to provision the wifi in the installed image.  This
commit lays the ground work for asking networking questions a future
time.

[ Issue: LINUXEXEC-8353 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>