u-boot-uenv: Switch to using devtype/devnum instead of mmc/mmcdev
The distro_bootcmd will automatically set the devtype and devnum so
that it will be possible to boot of more than one type of device.
This allows for full support of emulators, usb and mmc devices.

The fallback for backward compatibility will be to use mmc and mmcdev
until all the BSPs are converted to use distro_bootcmd.

[ Issue: LINUXEXEC-8353 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>