initramfs: Fix ima iversion to work with busybox
Busybox's mount can use i_version instead of iversion, plus it is not
needed at all when the ima init script has not been included in the
initramfs.

This has the side effect that some of the util-linux commands can be
removed from the initramfs.  This patch specifically removes:
      util-linux-mount, util-linux-setsid

The util-linux base was removed because it was not actually adding
anything to the busybox initramfs.

[ Issue: LINCD-2833 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>