bootfs.sh: Improve to work with rm_work enabled
When the rm_work is enabled some of the directories full of symlinks
are removed which are needed to run the bootfs.sh and wic tools.

All the binaries are still there as long as the "bitbake
build-sysroots" has been run one time.  This patch will automatically
run the built-sysroots if needed and then use the resulting binary
link tree.

[ Issue: LINCD-2833 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>