genimage: add default OCI config.json
Issue: LINUXEXEC-9410

Follow the OCI runtime specification, the container is paired with a
config.json. This configuration file contains metadata necessary to
implement standard operations against the container. This includes the
process to run, environment variables to inject, sandboxing features to
use, etc.
See Open Container Initiative Runtime Specification
(https://github.com/opencontainers/runtime-spec) to get more informaton.

This default config.json is just a starter file. Editing of it is
required to achieve desired results. This original file is generated by
oci-runtime-tools. See runtime-tools
(https://github.com/opencontainers/runtime-tools) to get more
information.

We can use runc command to create an instance of a container for a
bundle. A standard container bundle contains all the information needed
to load and run a container. This includes the following artifacts:
config.json: contains configuration data. This required file must reside
             in the root of the bundle directory and must be named
	     config.json.
container's root filesystem: the directory referenced by root.path, if
                             that property is set in config.json.

Here is an example to run an instance of wrlinux container:
  $ mkdir container-bundle
  $ cd container-bundle
  $ cp DEPLOYDIR/config.json .
  $ mkdir rootfs
  $ tar -C rootfs -xjf DEPLOYDIR/wrlinux-image-small.container.tar.bz2
  $ sudo runc run CONTAINER-ID

Where CONTAINER-ID is the name for the instance of the container.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>