bootfs.sh: fix '-a "console=XXX"' not work on installed boot
The orginal sed match "s#console=[^ ]* console=[^]*#${OSTREE_CONSOLE}#g"
is not correct, it may not have two consoles to match.

This patch is to remove old console first and then append new ones

The /^setenv instdef/ is used for the 1st installing boot and
the /^setenv bootargs/ is used for the latter installed boot.

Read extra console (-a "console=xxx") and set it to installed boot

Issue: LINCD-2990

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>