genimage: fix unstable installation of complementary packages
Issue: LINUXEXEC-9819

We are having problem that the installing complementary packages are
not stable. The root cause is that run_cmd from utils.py cannot get
all output. This is OK if the output is only used for logging purpose,
but it becomes a critical issue when the output is used as some input
data. So use subprocess.check_output instead. This ensures that the
complementary packages are always installed and the result is deterministic.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>