appsdk: support 'external-packages'
Issue: LINUXEXEC-9819

Add support of 'external-packages' for appsdk.

If a package is in 'external-packages', then the complementary packages
logic does not apply to it. This is because we should allow for different
package splitting policy from yocto's one.

So the package installation process is:
1) Install packages specified in 'packages' section
2) Install complementary packages
3) Install packages specified in 'external-packages'

So what if users want to also install the dev or dbg packages of external
packages? They should explicitly specify it. e.g.
"""
external-packages:
- MyApp
- MyApp-dev
- MyApp-dbg
"""
Again, we should not put any restriction on it, because it's possible that
MyApp has already got development files and debug files in it. The above
only serves as an example.

What if the specified external package has the same name with some official
package already installed?
In this case, we should warn the user and proceed with the installation process.
Why choose such behavior? Because there are too many possibilities. We should
just warn the user and let them handle it.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>