genimage: generate a real docker image
Previously, the container image (container-base-intel-x86-64.container.tar.bz2)
is not a real docker image, it is just a tarball of rootfs snapshot which is
used as the input of docker import. [1]

Although `docker import' could convert the rootfs tarball to a docker image,
but in order to run it, the user has to use Dockerfile with ADD + CMD to build
or docker run with extra command and args.

Previously, we use sloci-image [2] to generate a single-layer OCI image,
in this commit, we use skopeo [3] to convert oci image rootfs to docker-archive
which is a tarball of docker image which could be used as the input of docker
load.

In this commit
- Drop oci image tarball, which the docker image is sufficient
  for skopeo copy to upload;

- Keep oci image rootfs for runc to run;

[1] https://pspdfkit.com/blog/2019/docker-import-export-vs-load-save/
[2] https://github.com/jirutka/sloci-image.git
[3] https://github.com/containers/skopeo/

Issue: LINUXEXEC-10230

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>