init-ostree-install.sh: Check tty device existance
Before launching mttyexec check each tty device by sending it an echo
and checking the return code.

If the firwmare does not make ttyS0 available for example the
following shows what happens at runtime.

  bash-5.0# echo > /dev/ttyS0 2> /dev/null || echo ttyS0 unavailable
  ttyS0 unavailable

This will fix the problem where mttyexec was hanging because it asked
to attach to a device that did not exist.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>