u-boot-uenv: fix ostree image crash on RPi4
While u-boot is upgraded to 2020.07, the ostree image crashes
on RPi4, thanks to the help of Limeng <Meng.Li@windriver.com>,
we found that the issue is caused by the commit [1] of new u-boot.
It removes additional 4k space for fdt allocation.

As commit [1] said: There is no technical reason to add additional 4k
space for FDT...none should modified DTB used by U-Boot.

Then Limeng suggested to remove fdtcontroladdr which is used as dtb
address to workaround the limitation of [1]

[1] https://github.com/u-boot/u-boot/commit/b8fd54d62f92658cbd20ca051304e13eabf24ddd

Issue: LINCD-3607

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>