genimage: implement runtime functionality of System Definition
Add a tool sysdef.sh to implement the runtime functionality of the
System Definition. From the running of scripts to the touching of
stamp files the System Definition tool(s) will process the files
defined in the 'system' tag.

...
root@intel-x86-64:~# sysdef.sh -h
usage: sysdef.sh [-f] [-v] |run-once|run-on-upgrade|run-always [script1] [script2] [...]
       sysdef.sh [-f] [-v] run-all
       sysdef.sh [-v] list
           -f: ignore stamp, force to run
           -v: verbose
...

The System Definition tool(s) should be run automatically when
included in an image but should also be able to be run manually, to
assist with debugging for example.

A systemd service will ensure the tool(s) are run at a suitable time
during boot and the syslog should be captured for the tool and each of
the scripts it runs.

Issue: LINUXEXEC-7235

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>